Move HomeSection to separate file
